Refactor tests to one class and test constants to one file.
