+++ src/test/resources/testng.xml
@@ -29,8 +29,7 @@
             <class name="org.wso2.carbon.module.swiftiso20022.ISOMessageParserTests" />
             <class name="org.wso2.carbon.module.swiftiso20022.payloadfactory.ISO20022ToMT940PayloadFactoryTests" />
             <class name="org.wso2.carbon.module.swiftiso20022.JsonToMT103ValidatorTests"/>
-            <class name="org.wso2.carbon.module.swiftiso20022.MTParserTests"/>
-            <class name="org.wso2.carbon.module.swiftiso20022.mtmessageparser.MTParserTests" />
+            <class name="org.wso2.carbon.module.swiftiso20022.mt.parsers.MTParserTests" />
         </classes>
     </test>
 </suite>
